workshop2-pdf-mcp/pyproject.toml
[project]
name = "workshop2-pdf-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.1",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typhoon-ocr>=0.4.1",
]
